# Optimize inner dynamic programming loop in mass alignment