This commit adds additional check before calling `score` in the inner dynamic programming loop that avoids calling expensive mass tolerance check in some cases. Note, that due to reordering of the dynamic programming transitions behavior of the algorithm may change in case if many alignments with maximum score exist.

Also, for performance testing a `--no-parallel` flag added to `de-novo-align` example as sequential execution gives more stable execution time.

Performance testing was done with `--peptides rustyms/data/200305_HER_test_04_DENOVO_excerpt.csv --database examples/de-novo-align/database.fasta --out-path /tmp/out.csv --sequential` and the following results were obtained.

Without patch:
```
real	0m7,446s
user	0m7,430s
sys	0m0,015s
```

With patch:
```
real	0m3,631s
user	0m3,600s
sys	0m0,023s
```